@@ -8,11 +8,11 @@
         }
     },
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=8.0.2",
         "guzzlehttp/guzzle": "^7.4.5",
-        "jms/serializer": "^1.11 || ^2.0 || ^3.0",
-        "sensio/framework-extra-bundle": ">=3.0.2",
-        "symfony/framework-bundle": "~2.6 || ^3.0 || ^4.0 || ^5.0 || ^6.0"
+        "symfony/framework-bundle": "^5.4 || ^6.0",
+        "symfony/serializer": "^5.4 || ^6.0",
+        "symfony/property-access": "^5.4 || ^6.0"
     },
     "require-dev": {
     },
